Add files via upload
Simple linear regression is a machine learning method used to model the relationship between a single independent variable (input) and a single dependent variable (output) using a straight line. The goal is to find the "best-fit" line that minimizes the distance between the predicted values and the actual observed values. This best-fit line is defined by its slope and intercept. Here, we have used the wine quality dataset from UCI Machine Learning Repository.