Add files via upload
Multiple linear regression is a statistical method used to model the relationship between a dependent variable and two or more independent variables by fitting a linear equation to observed data. It aims to understand how changes in the independent variables affect the dependent variable, allowing for predictions of the dependent variable's value based on the independent variables. Here, I have used the startup dataset to analyze.