Add files via upload
The project demonstrates the implementation of Support Vector Regression (SVR) using the popular Position-Salary dataset. SVR is a type of supervised machine learning algorithm that is specifically designed for regression tasks, aiming to predict continuous target variables. It is an extension of the Support Vector Machine (SVM), which is primarily used for classification. The SVR model is particularly useful when dealing with non-linear relationships between features and target values, and this example helps illustrate how SVR can model such data effectively.